+++ src/cell_potts_model/EnergyWriter.java
@@ -3,9 +3,9 @@ package cell_potts_model;
 public class EnergyWriter extends DataWriter {
 
 	@Override
-	public void writeData(CellPottsModel model, double time) {
+	public void writeData(CellPottsModel model, int time) {
 		writer.println();		
-		writer.printf("%.8f %.8f", time, model.getTotalEnergy());
+		writer.printf("%.8f %.8f", (double) time, model.getTotalEnergy());
 	}
 
 }
